Break out test function from header into source file, rename.
