+++ tests/classes/cells3.cpp
@@ -1,9 +1,11 @@
 // SPDX-License-Identifier: GPL-3.0-or-later
 // Copyright (c) 2026 Team Dissolve and contributors
 
+#include "classes/configuration.h"
 #include "kernels/energy.h"
+#include "nodes/species.h"
 #include "templates/algorithms.h"
-#include "tests/graphData.h"
+#include "tests/testing.h"
 #include <gtest/gtest.h>
 
 namespace UnitTest
