+++ tests/classes/history.cpp
@@ -3,8 +3,8 @@
 
 #include "math/history.h"
 #include "nodes/number.h"
-#include "tests/graphData.h"
 #include "tests/testData.h"
+#include "tests/testing.h"
 #include <gtest/gtest.h>
 
 namespace UnitTest
