+++ tests/classes/structure.cpp
@@ -3,7 +3,7 @@
 
 #include "classes/structure.h"
 #include "nodes/calculateBonding.h"
-#include "tests/graphData.h"
+#include "tests/testing.h"
 #include <cmath>
 #include <gtest/gtest.h>
