+++ tests/ff/cos-n.cpp
@@ -3,6 +3,7 @@
 
 #include "tests/graphData.h"
 #include "tests/testData.h"
+#include "tests/testing.h"
 #include <gtest/gtest.h>
 
 namespace UnitTest
