+++ tests/gui/isotopologueSetModel.cpp
@@ -3,7 +3,8 @@
 
 #include "gui/models/isotopologueSetModel.h"
 #include "classes/isotopologueSet.h"
-#include "tests/graphData.h"
+#include "classes/species.h"
+#include "tests/testing.h"
 #include <gtest/gtest.h>
 
 namespace UnitTest
