+++ tests/nodes/angle.cpp
@@ -5,8 +5,8 @@
 #include "classes/speciesSites.h"
 #include "math/rangedVector3.h"
 #include "nodes/iterableGraph.h"
-#include "tests/graphData.h"
 #include "tests/testData.h"
+#include "tests/testing.h"
 #include <gtest/gtest.h>
 
 namespace UnitTest
