+++ tests/nodes/atomicMC.cpp
@@ -6,8 +6,8 @@
 #include "nodes/dissolve.h"
 #include "nodes/insert.h"
 #include "nodes/iterableGraph.h"
-#include "tests/graphData.h"
 #include "tests/testData.h"
+#include "tests/testing.h"
 #include <gtest/gtest.h>
 #include <vector>
