+++ tests/nodes/bragg.cpp
@@ -8,8 +8,8 @@
 #include "nodes/gr.h"
 #include "nodes/iterableGraph.h"
 #include "nodes/sq.h"
-#include "tests/graphData.h"
 #include "tests/testData.h"
+#include "tests/testing.h"
 #include <gtest/gtest.h>
 #include <memory>
