+++ tests/nodes/broadening.cpp
@@ -3,8 +3,10 @@
 
 #include "math/windowFunction.h"
 #include "nodes/gr.h"
-#include "tests/graphData.h"
+#include "nodes/neutronSQ.h"
+#include "nodes/sq.h"
 #include "tests/testData.h"
+#include "tests/testing.h"
 #include <gtest/gtest.h>
 
 namespace UnitTest
