+++ tests/nodes/dAngle.cpp
@@ -5,8 +5,8 @@
 #include "analyser/dataOperator2D.h"
 #include "nodes/importDLPUtilsSurface.h"
 #include "nodes/iterableGraph.h"
-#include "tests/graphData.h"
 #include "tests/testData.h"
+#include "tests/testing.h"
 #include <gtest/gtest.h>
 
 namespace UnitTest
