+++ tests/nodes/epsr.cpp
@@ -4,8 +4,12 @@
 #include "nodes/epsr.h"
 #include "classes/configuration.h"
 #include "main/dissolve.h"
-#include "tests/graphData.h"
+#include "nodes/gr.h"
+#include "nodes/neutronSQ.h"
+#include "nodes/sq.h"
+#include "nodes/xRaySQ.h"
 #include "tests/testData.h"
+#include "tests/testing.h"
 #include <gtest/gtest.h>
 #include <vector>
