+++ tests/nodes/gr.cpp
@@ -3,8 +3,8 @@
 
 #include "nodes/gr.h"
 #include "math/windowFunction.h"
-#include "tests/graphData.h"
 #include "tests/testData.h"
+#include "tests/testing.h"
 #include <gtest/gtest.h>
 
 namespace UnitTest
