+++ tests/nodes/histogramCN.cpp
@@ -3,8 +3,9 @@
 
 #include "nodes/histogramCN.h"
 #include "nodes/angle.h"
-#include "tests/graphData.h"
+#include "nodes/iterableGraph.h"
 #include "tests/testData.h"
+#include "tests/testing.h"
 #include <gtest/gtest.h>
 
 namespace UnitTest
