+++ tests/nodes/moleculeTorsion.cpp
@@ -3,8 +3,8 @@
 
 #include "nodes/moleculeTorsion.h"
 #include "nodes/iterableGraph.h"
-#include "tests/graphData.h"
 #include "tests/testData.h"
+#include "tests/testing.h"
 #include <gtest/gtest.h>
 
 namespace UnitTest
