+++ tests/nodes/neutronSQ.cpp
@@ -1,10 +1,12 @@
 // SPDX-License-Identifier: GPL-3.0-or-later
 // Copyright (c) 2026 Team Dissolve and contributors
 
+#include "nodes/neutronSQ.h"
 #include "math/windowFunction.h"
 #include "nodes/gr.h"
-#include "tests/graphData.h"
+#include "nodes/sq.h"
 #include "tests/testData.h"
+#include "tests/testing.h"
 #include <gtest/gtest.h>
 
 namespace UnitTest
