+++ tests/nodes/sdf.cpp
@@ -4,8 +4,8 @@
 #include "nodes/sdf.h"
 #include "nodes/importDLPUtilsPDens.h"
 #include "nodes/iterableGraph.h"
-#include "tests/graphData.h"
 #include "tests/testData.h"
+#include "tests/testing.h"
 #include <gtest/gtest.h>
 
 namespace UnitTest
