+++ tests/nodes/voxelDensity.cpp
@@ -3,8 +3,9 @@
 
 #include "nodes/voxelDensity.h"
 #include "data/atomicMasses.h"
-#include "tests/graphData.h"
+#include "nodes/configuration.h"
 #include "tests/testData.h"
+#include "tests/testing.h"
 #include <gtest/gtest.h>
 
 namespace UnitTest
