+++ tests/nodes/xRaySQ.cpp
@@ -3,8 +3,8 @@
 
 #include "nodes/xRaySQ.h"
 #include "nodes/gr.h"
-#include "tests/graphData.h"
 #include "tests/testData.h"
+#include "tests/testing.h"
 #include <gtest/gtest.h>
 
 namespace UnitTest
