+++ tests/workflows/phantomAtoms.cpp
@@ -2,8 +2,12 @@
 // Copyright (c) 2026 Team Dissolve and contributors
 
 #include "data/atomicMasses.h"
-#include "tests/graphData.h"
+#include "nodes/gr.h"
+#include "nodes/neutronSQ.h"
+#include "nodes/sq.h"
+#include "nodes/xRaySQ.h"
 #include "tests/testData.h"
+#include "tests/testing.h"
 #include <gtest/gtest.h>
 
 namespace UnitTest
